Split out interactions with source from service
