+++ common/src/main/scala/uk/ac/wellcome/platform/stacks/common/services/SierraService.scala
@@ -1,66 +1,31 @@
 package uk.ac.wellcome.platform.stacks.common.services
 
-import java.time.Instant
-
-import akka.actor.ActorSystem
-import akka.http.scaladsl.model.Uri
-import akka.http.scaladsl.model.Uri.Path
-import akka.http.scaladsl.model.headers.{Authorization, BasicHttpCredentials}
-import akka.stream.ActorMaterializer
 import uk.ac.wellcome.platform.stacks.common.models._
 
-import scala.concurrent.Future
+import scala.concurrent.{ExecutionContext, Future}
 
 class SierraService(
-    val baseUri: Uri = Uri(
-      "https://libsys.wellcomelibrary.org/iii/sierra-api"
-    ),
-    credentials: BasicHttpCredentials
-)(
-    implicit
-    val system: ActorSystem,
-    val mat: ActorMaterializer
-) extends AkkaClientService
-    with AkkaClientServiceGet
-    with AkkaClientServicePost
-    with AkkaClientTokenExchange {
+    sierraSource: SierraSource
+)(implicit ec: ExecutionContext) {
 
-  import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport._
-  import io.circe.generic.auto._
+  import SierraSource._
 
-  override val tokenPath = Path("v5/token")
-
-  // See https://sandbox.iii.com/iii/sierra-api/swagger/index.html#!/items
   def getItemStatus(sierraId: SierraItemIdentifier): Future[StacksItemStatus] =
-    for {
-      token <- getToken(credentials)
-      item <- get[SierraItemStub](
-        path = Path(s"v5/items/$sierraId"),
-        headers = List(Authorization(token))
-      )
+    sierraSource
+      .getSierraItemStub(sierraId)
+      .map(item => StacksItemStatus(item.status.code))
 
-    } yield StacksItemStatus(item.status.code)
-
-  // See https://sandbox.iii.com/iii/sierra-api/swagger/index.html#!/patrons
   def placeHold(
       userIdentifier: StacksUserIdentifier,
       sierraItemIdentifier: SierraItemIdentifier,
       itemLocation: StacksLocation
   ): Future[Unit] =
-    for {
-      token <- getToken(credentials)
-      _ <- post[SierraHoldRequestPostBody, String](
-        path = Path(s"v5/patrons/${userIdentifier.value}/holds/requests"),
-        body = Some(
-          SierraHoldRequestPostBody(
-            recordType = "i",
-            recordNumber = sierraItemIdentifier.value,
-            pickupLocation = itemLocation.id
-          )
-        ),
-        headers = List(Authorization(token))
+    sierraSource
+      .postHold(
+        userIdentifier,
+        sierraItemIdentifier,
+        itemLocation
       )
-    } yield ()
 
   protected def buildStacksHold(
       entry: SierraUserHoldsEntryStub
@@ -87,60 +52,16 @@ class SierraService(
     StacksHold(itemId, pickup, status)
   }
 
-  // See https://sandbox.iii.com/iii/sierra-api/swagger/index.html#!/patrons
   def getStacksUserHolds(
       userId: StacksUserIdentifier
   ): Future[StacksUserHolds[SierraItemIdentifier]] = {
-    for {
-      token <- getToken(credentials)
-
-      item <- get[SierraUserHoldsStub](
-        path = Path(s"v5/patrons/${userId.value}/holds"),
-        params = Map(
-          ("limit", "100"),
-          ("offset", "0")
-        ),
-        headers = List(Authorization(token))
-      )
-
-      holds = item.entries.map(buildStacksHold)
-
-    } yield StacksUserHolds(
-      userId = userId.value,
-      holds = holds
-    )
+    sierraSource
+      .getSierraUserHoldsStub(userId)
+      .map { hold =>
+        StacksUserHolds(
+          userId = userId.value,
+          holds = hold.entries.map(buildStacksHold)
+        )
+      }
   }
-
-  case class SierraUserHoldsPickupLocationStub(
-      code: String,
-      name: String
-  )
-  case class SierraUserHoldsStatusStub(
-      code: String,
-      name: String
-  )
-  case class SierraUserHoldsEntryStub(
-      id: String,
-      record: String,
-      pickupLocation: SierraUserHoldsPickupLocationStub,
-      pickupByDate: Option[Instant],
-      status: SierraUserHoldsStatusStub
-  )
-  case class SierraUserHoldsStub(
-      total: Long,
-      entries: List[SierraUserHoldsEntryStub]
-  )
-  case class SierraItemStatusStub(
-      code: String,
-      display: String
-  )
-  case class SierraItemStub(
-      id: String,
-      status: SierraItemStatusStub
-  )
-  case class SierraHoldRequestPostBody(
-      recordType: String,
-      recordNumber: Long,
-      pickupLocation: String
-  )
 }
